+++ src/main/java/com/devbliss/gpullr/util/http/ValuePairListFactory.java
@@ -1,0 +1,15 @@
+package com.devbliss.gpullr.util.http;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * Injectable class that creates new instances of {@link ValuePairList}s
+ */
+@Component
+public class ValuePairListFactory {
+
+  public ValuePairList getNewValuePairList(int initialCapacity) {
+    return ValuePairList.create(initialCapacity);
+  }
+
+}