+++ changelog.md
@@ -2,6 +2,8 @@
 ### Version 0.1.0-SNAPSHOT
 
  - [bug] fetching pull request events with updated repo url after repo renaming
+ - [feature] implemented GitHub OAuth login web flow
+ - [feature] added Fränkisch to list of valid languages
  - [feature] fair ranking algorithm
  - [feature] add allowed languages
  - [feature] no more pr ordering in backend / two order fields for assigned and unassigned prs
@@ -61,4 +63,4 @@
  - [feature] project setup
  - [initial] initial commit
 
--- Last change by: hschuetz <henning.schuetz@devbliss.com> Wed May 06 10:10:12 CEST 2015
+-- Last change by: Marco Reinwarth <marco.reinwarth@devbliss.com> Wed May 06 16:10:20 CEST 2015